src/loadtest/requirements.txt
locust
asciichartpy==1.5.25
tabulate==0.9.0
azure-identity==1.16.0
azure-monitor-query==1.3.0
azure-monitor-opentelemetry==1.3.0
